# apps/web/openapi/gateway.json
{
  "components": {
    "schemas": {
...
      "AssetResponse": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "height": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mediaType": {
            "title": "Mediatype",
            "type": "string"
          },
          "previewUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previewurl"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeneratedImageProvenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "sizeBytes": {
            "minimum": 0.0,
            "title": "Sizebytes",
            "type": "integer"
          },
          "sourceJobId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcejobid"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "enum": [
              "image",
              "video",
              "template"
            ],
            "title": "Type",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "createdAt",
          "mediaType",
          "sizeBytes"
        ],
        "title": "AssetResponse",
        "type": "object"
      },
...
      "GeneratedImageProvenance": {
        "additionalProperties": false,
        "properties": {
          "agentRunId": {
            "title": "Agentrunid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "outputIndex": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Outputindex",
            "type": "integer"
          },
          "promptSha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Promptsha256",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "providerResponseId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providerresponseid"
          },
          "purpose": {
            "enum": [
              "first-frame",
              "last-frame",
              "style-reference"
            ],
            "title": "Purpose",
            "type": "string"
          },
          "ratio": {
            "enum": [
              "21:9",
              "16:9",
              "4:3",
              "1:1",
              "3:4",
              "9:16"
            ],
            "title": "Ratio",
            "type": "string"
          },
          "safetyOutcome": {
            "const": "accepted",
            "default": "accepted",
            "title": "Safetyoutcome",
            "type": "string"
          },
          "sourceType": {
            "const": "agent-image",
            "default": "agent-image",
            "title": "Sourcetype",
            "type": "string"
          },
          "toolCallId": {
            "title": "Toolcallid",
            "type": "string"
          }
        },
        "required": [
          "agentRunId",
          "toolCallId",
          "outputIndex",
          "provider",
          "model",
          "promptSha256",
          "purpose",
          "ratio",
          "createdAt"
        ],
        "title": "GeneratedImageProvenance",
        "type": "object"
      },
      "GenerationDraft": {
        "additionalProperties": false,
        "properties": {
          "duration": {
            "default": 5,
            "maximum": 15.0,
            "minimum": 1.0,
            "title": "Duration",
            "type": "integer"
          },
          "enhancePrompt": {
            "default": false,
            "title": "Enhanceprompt",
            "type": "boolean"
          },
          "firstFrameAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstframeassetid"
          },
          "firstStrength": {
            "default": 1,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Firststrength",
            "type": "number"
          },
          "lastFrameAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastframeassetid"
          },
          "lastStrength": {
            "default": 1,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Laststrength",
            "type": "number"
          },
          "mode": {
            "const": "I2V",
            "default": "I2V",
            "title": "Mode",
            "type": "string"
          },
          "negativePrompt": {
            "default": "",
            "title": "Negativeprompt",
            "type": "string"
          },
          "offload": {
            "default": "none",
            "enum": [
              "none",
              "cpu"
            ],
            "title": "Offload",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileTier",
            "default": "fast"
          },
          "prompt": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "quantization": {
            "default": "fp8-cast",
            "enum": [
              "fp8-cast",
              "none"
            ],
            "title": "Quantization",
            "type": "string"
          },
          "queue": {
            "$ref": "#/components/schemas/QueueTier",
            "default": "fast"
          },
          "ratio": {
            "default": "16:9",
            "enum": [
              "21:9",
              "16:9",
              "4:3",
              "1:1",
              "3:4",
              "9:16"
            ],
            "title": "Ratio",
            "type": "string"
          },
          "resolution": {
            "default": "720p",
            "enum": [
              "720p",
              "1080p"
            ],
            "title": "Resolution",
            "type": "string"
          },
          "seed": {
            "default": 42,
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "GenerationDraft",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobCreate": {
        "additionalProperties": false,
        "properties": {
          "computeSessionId": {
            "title": "Computesessionid",
            "type": "string"
          },
          "conversationId": {
            "title": "Conversationid",
            "type": "string"
          },
          "draft": {
            "$ref": "#/components/schemas/GenerationDraft"
          }
        },
        "required": [
          "conversationId",
          "computeSessionId",
          "draft"
        ],
        "title": "JobCreate",
        "type": "object"
      },
...
      "JobResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "default": 1,
            "minimum": 1.0,
            "title": "Attempt",
            "type": "integer"
          },
          "computeSessionId": {
            "title": "Computesessionid",
            "type": "string"
          },
          "conversationId": {
            "title": "Conversationid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "currentStep": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentstep"
          },
          "draft": {
            "$ref": "#/components/schemas/GenerationDraft"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "gpu": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GpuAssignment"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobOutput"
              },
              {
                "type": "null"
              }
            ]
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "profileId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profileid"
          },
          "progress": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "integer"
          },
          "queuePosition": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queueposition"
          },
          "stage": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "totalSteps": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalsteps"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "warmStart": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warmstart"
          }
        },
        "required": [
          "id",
          "conversationId",
          "computeSessionId",
          "createdAt",
          "updatedAt",
          "stage",
          "progress",
          "draft"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "draft",
          "uploaded",
          "queued",
          "assigned",
          "loading_model",
          "preparing",
          "generating",
          "encoding",
          "cancel_requested",
          "succeeded",
          "cancelled",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Oneiroi Studio Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/assets": {
      "get": {
        "operationId": "list_assets_v1_assets_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetResponse"
                  },
                  "title": "Response List Assets V1 Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Assets",
        "tags": [
          "assets"
        ]
      }
    },
    "/v1/assets/{asset_id}": {
      "delete": {
        "operationId": "delete_asset_v1_assets__asset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Asset",
        "tags": [
          "assets"
        ]
      }
    },
    "/v1/assets/{asset_id}/file": {
      "get": {
        "operationId": "get_asset_file_v1_assets__asset_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asset File",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/v1/jobs": {
      "get": {
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  },
                  "title": "Response List Jobs V1 Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/i2v": {
      "post": {
        "operationId": "create_job_v1_jobs_i2v_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "operationId": "job_events_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/file": {
      "get": {
        "operationId": "get_job_file_v1_jobs__job_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job File",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/manifest": {
      "get": {
        "operationId": "get_job_manifest_v1_jobs__job_id__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Manifest",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_job_v1_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}